+++ src/core/properties/propertyownerobserver.cpp
@@ -34,30 +34,32 @@
 
 namespace inviwo {
 
-void PropertyOwnerObservable::notifyObserversWillAddProperty(Property* property, int index) const {
+void PropertyOwnerObservable::notifyObserversWillAddProperty(Property* property,
+                                                             size_t index) const {
     for (ObserverSet::reverse_iterator it = observers_->rbegin(); it != observers_->rend(); ++it) {
         static_cast<PropertyOwnerObserver*>(*it)->onWillAddProperty(property, index);
     }
 }
 
-void PropertyOwnerObservable::notifyObserversDidAddProperty(Property* property, int index) const {
+void PropertyOwnerObservable::notifyObserversDidAddProperty(Property* property,
+                                                            size_t index) const {
     for (ObserverSet::reverse_iterator it = observers_->rbegin(); it != observers_->rend(); ++it) {
         static_cast<PropertyOwnerObserver*>(*it)->onDidAddProperty(property, index);
     }
 }
 
-void PropertyOwnerObservable::notifyObserversWillRemoveProperty(Property* property, int index) const {
+void PropertyOwnerObservable::notifyObserversWillRemoveProperty(Property* property,
+                                                                size_t index) const {
     for (ObserverSet::reverse_iterator it = observers_->rbegin(); it != observers_->rend(); ++it) {
         static_cast<PropertyOwnerObserver*>(*it)->onWillRemoveProperty(property, index);
     }
 }
 
-void PropertyOwnerObservable::notifyObserversDidRemoveProperty(Property* property, int index) const {
+void PropertyOwnerObservable::notifyObserversDidRemoveProperty(Property* property,
+                                                               size_t index) const {
     for (ObserverSet::reverse_iterator it = observers_->rbegin(); it != observers_->rend(); ++it) {
         static_cast<PropertyOwnerObserver*>(*it)->onDidRemoveProperty(property, index);
     }
 }
 
-
-} // namespace
-
+}  // namespace
